+++ src/app/components/heimdall-roles/heimdall-roles.component.ts
@@ -16,6 +16,8 @@ export class HeimdallRolesComponent implements OnInit {
   displayedColumns: string[] = [ 'Roles'];
   allRoles;
   newRole="";
+  isAuthorized = this.service.authorized;
+  
   ngOnInit(): void {
     this.getAllRoles();
   }
@@ -40,12 +42,14 @@ export class HeimdallRolesComponent implements OnInit {
   }
 
   getAllRoles(){
+    if(this.service.authorized)
    this.service.getAllRoles().subscribe(data=>{
      this.allRoles = data;
    })
   }
   
   addRole(){
+    if(this.service.authorized)
    this.service.addRole(this.newRole).subscribe(data =>{
      this.getAllRoles();
    },error=>{
@@ -54,6 +58,7 @@ export class HeimdallRolesComponent implements OnInit {
   }
 
   deleteRole(role){
+    if(this.service.authorized)
     this.service.deleteRole(role).subscribe(data =>{
       this.getAllRoles()
     },error=>{
